Move DataGenerator under kas::view::clerk
